#7 change type from int to byte

+++ src/algoBhaiya.ReportBook.Mobile/algoBhaiya.ReportBook.Core/Entities/MonthlyTarget.cs
@@ -10,7 +10,7 @@ namespace algoBhaiya.ReportBook.Core.Entities
         public byte UserId { get; set; }
         public int FieldTemplateId { get; set; }
 
-        public int Month { get; set; }
+        public byte Month { get; set; }
         public int Year { get; set; }
 
         public string TargetValue { get; set; }
